\name{concerto5-package}
\alias{concerto5-package}
\docType{package}
\title{
    Package for Concerto platform test logic R code.
}
\description{
    Contains all required functions by Concerto platform.
}
\details{
\tabular{ll}{
    Package: \tab concerto5\cr
    Type: \tab Package\cr
    Version: \tab 1.33.0\cr
    Date: \tab 2021-11-08\cr
    License: \tab Apache-2.0\cr
}

}
\author{
    Przemyslaw Lis
    Maintainer: Przemyslaw Lis <pl362@cam.ac.uk>
}
\references{
}
\keyword{package}
\seealso{
}
\examples{
}
