\name{concerto.file.getPath}
\alias{concerto.file.getPath}
\title{
    Public path
}
\description{
    Gets file public path.
}
\usage{
    concerto.file.getPath(filename)
}
\arguments{
    \item{filename}{
        File name
    }
}
\details{
}
\value{
    Public path of given file name
}
\references{
}
\author{
    Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
    plot_path <- concerto.file.getPath("plot.png")
}

\keyword{table}
\keyword{database}
