\name{concerto.file.getPublicPath}
\alias{concerto.file.getPublicPath}
\title{
Public path
}
\description{
Gets file public path.
}
\usage{
concerto.file.getPublicPath(filename)
}
\arguments{
  \item{filename}{
File name
}
}
\details{
}
\value{
Public path of given file name
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
plot_path <- concerto.file.getPublicPath("plot.png")
}

\keyword{ table }
\keyword{ database }
