\name{leap.list.merge}
\alias{leap.list.merge}
\title{
    Merge two lists.
}
\description{
    Merges two lists
}
\usage{
    leap.list.merge(list1, list2)
}
\arguments{
    \item{list1}{
        First list
    }
    \item{list2}{
        Second list
    }
}
\details{
}
\value{
    Merged list
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    res <- leap.list.merge(list1, list2)
}

\keyword{list}
