\name{concerto.log}
\alias{concerto.log}
\title{
    Logs
}
\description{
    Logs message
}
\usage{
    concerto.log(obj)
}
\arguments{
    \item{obj}{
        Object to log
    }
    \item{title}{
        Object description
    }
}
\details{
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    concerto.log("success")
}

\keyword{log}
