\name{concerto.session.restoreFlow}
\alias{concerto.session.restoreFlow}
\title{
Restores flow based session
}
\description{
Restore flow based test session.
}
\usage{
concerto.session.restoreFlow(sessionHash)
}
\arguments{
  \item{sessionHash}{
Test session hash.
}
}
\details{
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
concerto.session.restoreFlow("414c9ff4bc7ff0b94a7348b38777fdfd45c03786")
}
\keyword{ session }
