\name{concerto.trans}
\alias{concerto.trans}
\alias{c.trans}
\title{
Translations
}
\description{
Translate entry using global dictionary
}
\usage{
concerto.trans(key)
}
\arguments{
  \item{key}{
Entry key
}
}
\details{
}
\value{
Translated entry
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
text <- concerto.trans("text")
}

\keyword{ variable }
