\name{concerto.var.get}
\alias{concerto.var.get}
\title{
Global variables
}
\description{
Gets global variable value
}
\usage{
concerto.var.get(name)
}
\arguments{
  \item{name}{
Variable name
}
}
\details{
}
\value{
Value of global variable or NULL if not set
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
score <- concerto.var.get("score")
}

\keyword{ variable }
