\name{concerto.var.getDynamicInputs}
\alias{concerto.var.getDynamicInputs}
\alias{c.getDynamicInputs}
\title{
Dynamic inputs
}
\description{
Gets vector of names of dynamic inputs in current context
}
\usage{
concerto.var.getDynamicInputs()
}
\arguments{
}
\details{
}
\value{
Vector of dynamic input names
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
names <- concerto.var.getDynamicInputs()
}

\keyword{ variable }
