\name{leap5-package}
\alias{leap5-package}
\docType{package}
\title{
    Package for LEAP Scores test logic R code.
}
\description{
    Contains all required functions by LEAP Scores.
}
\details{
\tabular{ll}{
    Package: \tab leap5\cr
    Type: \tab Package\cr
    Version: \tab 0.35.0\cr
    Date: \tab 2022-06-24\cr
    License: \tab Apache-2.0\cr
}

}
\author{
    Kooky Whiz
    Maintainer: Kooky Whiz <pl362@cam.ac.uk>
}
\references{
}
\keyword{package}
\seealso{
}
\examples{
}
