\name{concerto.event.fire}
\alias{concerto.event.fire}
\title{
    Events
}
\description{
    Firing events.
}
\usage{
    concerto.event.fire(name)
}
\arguments{
    \item{name}{
        Event name
    }
}
\details{
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    concerto.event.fire("onTemplateSubmit")
}

\keyword{events}
