\name{leap.file.getPath}
\alias{leap.file.getPath}
\title{
    Public path
}
\description{
    Gets file public path.
}
\usage{
    leap.file.getPath(filename)
}
\arguments{
    \item{filename}{
        File name
    }
}
\details{
}
\value{
    Public path of given file name
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    plot_path <- leap.file.getPath("plot.png")
}

\keyword{table}
\keyword{database}
