\name{concerto.file.getUrl}
\alias{concerto.file.getUrl}
\title{
Files URL
}
\description{
Gets URL of filename in public directory.
}
\usage{
concerto.file.getUrl(filename, noCache=F)
}
\arguments{
  \item{filename}{
File name
}
  \item{noCache}{
Should not try to use cache?
}
}
\details{
}
\value{
URL of filename in public directory
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
plot_url <- concerto.file.getUrl("plot.png")
}

\keyword{ table }
\keyword{ database }
