\name{leap.file.getUrl}
\alias{leap.file.getUrl}
\title{
    Files URL
}
\description{
    Gets URL of filename in public directory.
}
\usage{
    leap.file.getUrl(filename, noCache=F)
}
\arguments{
    \item{filename}{
        File name
    }
    \item{noCache}{
        Should not try to use cache?
    }
}
\details{
}
\value{
    URL of filename in public directory
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    plot_url <- leap.file.getUrl("plot.png")
}

\keyword{table}
\keyword{database}
