\name{concerto.list.merge}
\alias{concerto.list.merge}
\title{
Merge two lists.
}
\description{
Merges two lists
}
\usage{
concerto.list.merge(list1, list2)
}
\arguments{
  \item{list1}{
First list
}
  \item{list2}{
Second list
}
}
\details{
}
\value{
Merged list
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
res <- concerto.list.merge(list1, list2)
}

\keyword{ list }
