\name{concerto.table.lastInsertId}
\alias{concerto.table.lastInsertId}
\title{
Last insert id
}
\description{
Gets last inserted record id
}
\usage{
concerto.table.lastInsertId()
}
\details{
}
\value{
Last insert id
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
lid <- concerto.table.lastInsertId()
}

\keyword{ table }
\keyword{ database }
