\name{leap.table.lastInsertId}
\alias{leap.table.lastInsertId}
\title{
    Last insert id
}
\description{
    Gets last inserted record id
}
\usage{
    leap.table.lastInsertId(connection)
}
\arguments{
    \item{connection}{
        Connection to use.
    }
}
\details{
}
\value{
    Last insert id
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    lid <- leap.table.lastInsertId()
}

\keyword{table}
\keyword{database}
