\name{leap.trans}
\alias{leap.trans}
\alias{c.trans}
\title{
    Translations
}
\description{
    Translate entry using global dictionary
}
\usage{
    leap.trans(key)
}
\arguments{
    \item{key}{
        Entry key
    }
}
\details{
}
\value{
    Translated entry
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    text <- leap.trans("text")
}

\keyword{translation}
