\name{concerto.var.set}
\alias{concerto.var.set}
\title{
Promoted variables
}
\description{
Sets promoted variable value
}
\usage{
concerto.var.set(name, value)
}
\arguments{
  \item{name}{
Variable name
}
  \item{value}{
Variable value
}
}
\details{
}
\value{
Value of promoted variable
}
\references{
}
\author{
Przemyslaw Lis
}
\note{
}

\seealso{
}
\examples{
concerto.var.set("score", 100)
}

\keyword{ variable }
