\name{leap.var.set}
\alias{leap.var.set}
\alias{c.set}
\title{
    Global variables
}
\description{
    Sets global variable value
}
\usage{
    leap.var.set(name, value)
}
\arguments{
    \item{name}{
        Variable name
    }
    \item{value}{
        Variable value
    }
    \item{global}{
        Logical: TRUE to set value in top most position
    }
    \item{flowIndexOffset}{
        Numerical: flow index offset
    }
    \item{flowIndex}{
        Numerical: flow index
    }
}
\details{
}
\value{
    Value of global variable
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    leap.var.set("score", 100)
}

\keyword{variable}
